feat: drap afunc
