+++ src/xfp/a/functions/__init__.py
@@ -1,0 +1,53 @@
+from xfp.a.functions.utils import (
+    AF0,
+    AF1,
+    AF2,
+    AF3,
+    AF4,
+    AF5,
+    AF6,
+    AF7,
+    AF8,
+    AF9,
+    AF10,
+    AF11,
+    AF12,
+    AF13,
+    AF14,
+    AF15,
+    AF16,
+    AF17,
+    AF18,
+    AF19,
+    AF20,
+    AF21,
+    AF22,
+)
+from xfp.a.functions.afunc import AFunc
+
+__all__ = [
+    "AF0",
+    "AF1",
+    "AF2",
+    "AF3",
+    "AF4",
+    "AF5",
+    "AF6",
+    "AF7",
+    "AF8",
+    "AF9",
+    "AF10",
+    "AF11",
+    "AF12",
+    "AF13",
+    "AF14",
+    "AF15",
+    "AF16",
+    "AF17",
+    "AF18",
+    "AF19",
+    "AF20",
+    "AF21",
+    "AF22",
+    "AFunc",
+]
