+++ src/xfp/functions/__init__.py
@@ -1,152 +1,12 @@
-from typing import Callable
-from inspect import signature
-from functools import partial
-from typing import TypeVar, ParamSpec, cast, Concatenate, Any
+from xfp.functions.utils import *  # noqa: F403
 from xfp.functions.xfunc import XFunc
 
-__all__ = ["XFunc"]
-
-A = TypeVar("A")
-P = ParamSpec("P")
-T = TypeVar("T")
-Self = TypeVar("Self")
-
-
-def id(e: Any) -> Any:
-    return e
-
-
 # fmt: off
-type F0[Out] =                                                                                                                                                                         Callable[[], Out]
-type F1[**In, Out] =                                                                                                                                                                   Callable[In, Out]
-type F2[**In1, **In2, Out] =                                                                                                                                                           F1[In1, F1[In2, Out]]
-type F3[**In1, **In2, **In3, Out] =                                                                                                                                                    F1[In1, F2[In2, In3, Out]]
-type F4[**In1, **In2, **In3, **In4, Out] =                                                                                                                                             F1[In1, F3[In2, In3, In4, Out]]
-type F5[**In1, **In2, **In3, **In4, **In5, Out] =                                                                                                                                      F1[In1, F4[In2, In3, In4, In5, Out]]
-type F6[**In1, **In2, **In3, **In4, **In5, **In6, Out] =                                                                                                                               F1[In1, F5[In2, In3, In4, In5, In6, Out]]
-type F7[**In1, **In2, **In3, **In4, **In5, **In6, **In7, Out] =                                                                                                                        F1[In1, F6[In2, In3, In4, In5, In6, In7, Out]]
-type F8[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, Out] =                                                                                                                 F1[In1, F7[In2, In3, In4, In5, In6, In7, In8, Out]]
-type F9[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, Out] =                                                                                                          F1[In1, F8[In2, In3, In4, In5, In6, In7, In8, In9, Out]]
-type F10[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, Out] =                                                                                                 F1[In1, F9[In2, In3, In4, In5, In6, In7, In8, In9, In10, Out]]
-type F11[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, Out] =                                                                                         F1[In1, F10[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out]]
-type F12[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, Out] =                                                                                 F1[In1, F11[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out]]
-type F13[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, Out] =                                                                         F1[In1, F12[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out]]
-type F14[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, Out] =                                                                 F1[In1, F13[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out]]
-type F15[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, Out] =                                                         F1[In1, F14[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out]]
-type F16[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, Out] =                                                 F1[In1, F15[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out]]
-type F17[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, Out] =                                         F1[In1, F16[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out]]
-type F18[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, Out] =                                 F1[In1, F17[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out]]
-type F19[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, Out] =                         F1[In1, F18[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out]]
-type F20[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, Out] =                 F1[In1, F19[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out]]
-type F21[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, Out] =         F1[In1, F20[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out]]
-type F22[**In1, **In2, **In3, **In4, **In5, **In6, **In7, **In8, **In9, **In10, **In11, **In12, **In13, **In14, **In15, **In16, **In17, **In18, **In19, **In20, **In21, **In22, Out] = F1[In1, F21[In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out]]
-# fmt: on
-
-
-def _curry(f: F1[Concatenate[A, P], T]) -> F1[[A], Any]:
-    if len(signature(f).parameters) == 1:
-        return cast(F1[[A], T], f)
-    else:
-
-        def g(x: A) -> F1:
-            return curry(cast(F1, partial(f, x)))
-
-        return g
-
-
-def _curry_method(f: F1[Concatenate[Self, A, P], T]) -> F1[[Self, A], Any]:
-    if len(signature(f).parameters) == 2:
-        return cast(F1[[Self, A], T], f)
-    else:
-
-        def g(self: Self, x: A) -> F1:
-            return curry(cast(F1, partial(f, self, x)))
-
-        return g
-
-
-def _tupled[Out](f: F1[..., Out]) -> F1[[tuple], Out]:
-    return lambda args: f(*args)
-
-
-# fmt: off
-def curry(f: F1[Concatenate[A, P], T]) -> F1[[A], Any]: return _curry(f)
-def curry0[Out](f: F1[[], Out])                                                                                                                                                                                                                                                      -> F0[Out]: return f
-def curry1[In1, Out](f: F1[[In1], Out])                                                                                                                                                                                                                                              -> F1[[In1], Out]: return _curry(f)
-def curry2[In1, In2, Out](f: F1[[In1, In2], Out])                                                                                                                                                                                                                                    -> F2[[In1], [In2], Out]: return _curry(f)
-def curry3[In1, In2, In3, Out](f: F1[[In1, In2, In3], Out])                                                                                                                                                                                                                          -> F3[[In1], [In2], [In3], Out]: return _curry(f)
-def curry4[In1, In2, In3, In4, Out](f: F1[[In1, In2, In3, In4], Out])                                                                                                                                                                                                                -> F4[[In1], [In2], [In3], [In4], Out]: return _curry(f)
-def curry5[In1, In2, In3, In4, In5, Out](f: F1[[In1, In2, In3, In4, In5], Out])                                                                                                                                                                                                      -> F5[[In1], [In2], [In3], [In4], [In5], Out]: return _curry(f)
-def curry6[In1, In2, In3, In4, In5, In6, Out](f: F1[[In1, In2, In3, In4, In5, In6], Out])                                                                                                                                                                                            -> F6[[In1], [In2], [In3], [In4], [In5], [In6], Out]: return _curry(f)
-def curry7[In1, In2, In3, In4, In5, In6, In7, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7], Out])                                                                                                                                                                                  -> F7[[In1], [In2], [In3], [In4], [In5], [In6], [In7], Out]: return _curry(f)
-def curry8[In1, In2, In3, In4, In5, In6, In7, In8, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8], Out])                                                                                                                                                                        -> F8[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], Out]: return _curry(f)
-def curry9[In1, In2, In3, In4, In5, In6, In7, In8, In9, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9], Out])                                                                                                                                                              -> F9[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], Out]: return _curry(f)
-def curry10[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10], Out])                                                                                                                                                 -> F10[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], Out]: return _curry(f)
-def curry11[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11], Out])                                                                                                                                     -> F11[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], Out]: return _curry(f)
-def curry12[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12], Out])                                                                                                                         -> F12[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], Out]: return _curry(f)
-def curry13[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13], Out])                                                                                                             -> F13[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], Out]: return _curry(f)
-def curry14[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14], Out])                                                                                                 -> F14[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], Out]: return _curry(f)
-def curry15[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15], Out])                                                                                     -> F15[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], Out]: return _curry(f)
-def curry16[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16], Out])                                                                         -> F16[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], Out]: return _curry(f)
-def curry17[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17], Out])                                                             -> F17[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], Out]: return _curry(f)
-def curry18[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18], Out])                                                 -> F18[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], Out]: return _curry(f)
-def curry19[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19], Out])                                     -> F19[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], Out]: return _curry(f)
-def curry20[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20], Out])                         -> F20[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], [In20], Out]: return _curry(f)
-def curry21[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21], Out])             -> F21[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], [In20], [In21], Out]: return _curry(f)
-def curry22[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22], Out]) -> F22[[In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], [In20], [In21], [In22], Out]: return _curry(f)
-# fmt: on
-
-
-# fmt: off
-def curry_method(f: F1[Concatenate[Self, A, P], T]) -> F1[[Self, A], Any]: return _curry_method(f)
-def curry_method0[Self, Out](f: F1[[Self], Out])                                                                                                                                                                                                                                                        -> F1[[Self], Out]: return f
-def curry_method1[Self, In1, Out](f: F1[[Self, In1], Out])                                                                                                                                                                                                                                              -> F1[[Self, In1], Out]: return _curry_method(f)
-def curry_method2[Self, In1, In2, Out](f: F1[[Self, In1, In2], Out])                                                                                                                                                                                                                                    -> F2[[Self, In1], [In2], Out]: return _curry_method(f)
-def curry_method3[Self, In1, In2, In3, Out](f: F1[[Self, In1, In2, In3], Out])                                                                                                                                                                                                                          -> F3[[Self, In1], [In2], [In3], Out]: return _curry_method(f)
-def curry_method4[Self, In1, In2, In3, In4, Out](f: F1[[Self, In1, In2, In3, In4], Out])                                                                                                                                                                                                                -> F4[[Self, In1], [In2], [In3], [In4], Out]: return _curry_method(f)
-def curry_method5[Self, In1, In2, In3, In4, In5, Out](f: F1[[Self, In1, In2, In3, In4, In5], Out])                                                                                                                                                                                                      -> F5[[Self, In1], [In2], [In3], [In4], [In5], Out]: return _curry_method(f)
-def curry_method6[Self, In1, In2, In3, In4, In5, In6, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6], Out])                                                                                                                                                                                            -> F6[[Self, In1], [In2], [In3], [In4], [In5], [In6], Out]: return _curry_method(f)
-def curry_method7[Self, In1, In2, In3, In4, In5, In6, In7, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7], Out])                                                                                                                                                                                  -> F7[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], Out]: return _curry_method(f)
-def curry_method8[Self, In1, In2, In3, In4, In5, In6, In7, In8, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8], Out])                                                                                                                                                                        -> F8[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], Out]: return _curry_method(f)
-def curry_method9[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9], Out])                                                                                                                                                              -> F9[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], Out]: return _curry_method(f)
-def curry_method10[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10], Out])                                                                                                                                                 -> F10[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], Out]: return _curry_method(f)
-def curry_method11[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11], Out])                                                                                                                                     -> F11[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], Out]: return _curry_method(f)
-def curry_method12[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12], Out])                                                                                                                         -> F12[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], Out]: return _curry_method(f)
-def curry_method13[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13], Out])                                                                                                             -> F13[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], Out]: return _curry_method(f)
-def curry_method14[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14], Out])                                                                                                 -> F14[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], Out]: return _curry_method(f)
-def curry_method15[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15], Out])                                                                                     -> F15[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], Out]: return _curry_method(f)
-def curry_method16[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16], Out])                                                                         -> F16[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], Out]: return _curry_method(f)
-def curry_method17[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17], Out])                                                             -> F17[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], Out]: return _curry_method(f)
-def curry_method18[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18], Out])                                                 -> F18[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], Out]: return _curry_method(f)
-def curry_method19[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19], Out])                                     -> F19[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], Out]: return _curry_method(f)
-def curry_method20[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20], Out])                         -> F20[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], [In20], Out]: return _curry_method(f)
-def curry_method21[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21], Out])             -> F21[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], [In20], [In21], Out]: return _curry_method(f)
-def curry_method22[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out](f: F1[[Self, In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22], Out]) -> F22[[Self, In1], [In2], [In3], [In4], [In5], [In6], [In7], [In8], [In9], [In10], [In11], [In12], [In13], [In14], [In15], [In16], [In17], [In18], [In19], [In20], [In21], [In22], Out]: return _curry_method(f)
-# fmt: on
-
-
-# fmt: off
-def tupled[Out](f: F1[..., Out])                                                                                                                                                                                                                                                      -> F1[[tuple], Out]: return _tupled(f) 
-def tupled1[In1, Out](f: F1[[In1], Out])                                                                                                                                                                                                                                              -> F1[[tuple[In1]], Out]: return _tupled(f)
-def tupled2[In1, In2, Out](f: F1[[In1, In2], Out])                                                                                                                                                                                                                                    -> F1[[tuple[In1, In2]], Out]: return _tupled(f)
-def tupled3[In1, In2, In3, Out](f: F1[[In1, In2, In3], Out])                                                                                                                                                                                                                          -> F1[[tuple[In1, In2, In3]], Out]: return _tupled(f)
-def tupled4[In1, In2, In3, In4, Out](f: F1[[In1, In2, In3, In4], Out])                                                                                                                                                                                                                -> F1[[tuple[In1, In2, In3, In4]], Out]: return _tupled(f)
-def tupled5[In1, In2, In3, In4, In5, Out](f: F1[[In1, In2, In3, In4, In5], Out])                                                                                                                                                                                                      -> F1[[tuple[In1, In2, In3, In4, In5]], Out]: return _tupled(f)
-def tupled6[In1, In2, In3, In4, In5, In6, Out](f: F1[[In1, In2, In3, In4, In5, In6], Out])                                                                                                                                                                                            -> F1[[tuple[In1, In2, In3, In4, In5, In6]], Out]: return _tupled(f)
-def tupled7[In1, In2, In3, In4, In5, In6, In7, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7], Out])                                                                                                                                                                                  -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7]], Out]: return _tupled(f)
-def tupled8[In1, In2, In3, In4, In5, In6, In7, In8, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8], Out])                                                                                                                                                                        -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8]], Out]: return _tupled(f)
-def tupled9[In1, In2, In3, In4, In5, In6, In7, In8, In9, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9], Out])                                                                                                                                                              -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9]], Out]: return _tupled(f)
-def tupled10[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10], Out])                                                                                                                                                 -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10]], Out]: return _tupled(f)
-def tupled11[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11], Out])                                                                                                                                     -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11]], Out]: return _tupled(f)
-def tupled12[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12], Out])                                                                                                                         -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12]], Out]: return _tupled(f)
-def tupled13[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13], Out])                                                                                                             -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13]], Out]: return _tupled(f)
-def tupled14[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14], Out])                                                                                                 -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14]], Out]: return _tupled(f)
-def tupled15[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15], Out])                                                                                     -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15]], Out]: return _tupled(f)
-def tupled16[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16], Out])                                                                         -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16]], Out]: return _tupled(f)
-def tupled17[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17], Out])                                                             -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17]], Out]: return _tupled(f)
-def tupled18[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18], Out])                                                 -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18]], Out]: return _tupled(f)
-def tupled19[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19], Out])                                     -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19]], Out]: return _tupled(f)
-def tupled20[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20], Out])                         -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20]], Out]: return _tupled(f)
-def tupled21[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21], Out])             -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21]], Out]: return _tupled(f)
-def tupled22[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22, Out](f: F1[[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22], Out]) -> F1[[tuple[In1, In2, In3, In4, In5, In6, In7, In8, In9, In10, In11, In12, In13, In14, In15, In16, In17, In18, In19, In20, In21, In22]], Out]: return _tupled(f)
+__all__ = [
+    "F0", "F1", "F2", "F3", "F4", "F5", "F6", "F7", "F8", "F9", "F10", "F11", "F12", "F13", "F14", "F15", "F16", "F17", "F18", "F19", "F20", "F21", "F22",  # noqa: F405
+    "curry", "curry0", "curry1", "curry2", "curry3", "curry4", "curry5", "curry6", "curry7", "curry8", "curry9", "curry10", "curry11", "curry12", "curry13", "curry14", "curry15", "curry16", "curry17", "curry18", "curry19", "curry20", "curry21", "curry22",  # noqa: F405
+    "curry_method", "curry_method0", "curry_method1", "curry_method2", "curry_method3", "curry_method4", "curry_method5", "curry_method6", "curry_method7", "curry_method8", "curry_method9", "curry_method10", "curry_method11", "curry_method12", "curry_method13", "curry_method14", "curry_method15", "curry_method16", "curry_method17", "curry_method18", "curry_method19", "curry_method20", "curry_method21", "curry_method22",  # noqa: F405
+    "tupled", "tupled1", "tupled2", "tupled3", "tupled4", "tupled5", "tupled6", "tupled7", "tupled8", "tupled9", "tupled10", "tupled11", "tupled12", "tupled13", "tupled14", "tupled15", "tupled16", "tupled17", "tupled18", "tupled19", "tupled20", "tupled21", "tupled22",  # noqa: F405
+    "XFunc"
+]
 # fmt: on
